1. Pull state out of SpotlightContainer
2. Move releaseContainer in SpotlightContainer
3. Move stateFromProps in useSpotlightContainer

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)
